Move Tooltip component to common dir
