+++ assets/js/components/ChecksSelection/ChecksSelectionHeader.jsx
@@ -3,7 +3,7 @@ import { EOS_PLAY_CIRCLE } from 'eos-icons-react';
 import classNames from 'classnames';
 
 import Button from '@common/Button';
-import Tooltip from '@components/Tooltip';
+import Tooltip from '@common/Tooltip';
 
 import { canStartExecution } from '@components/ChecksSelection';
