+++ assets/js/components/SapSystemDetails/tableConfigs.jsx
@@ -4,7 +4,7 @@ import classNames from 'classnames';
 import HostLink from '@components/HostLink';
 import ProviderLabel from '@components/ProviderLabel';
 import CleanUpButton from '@components/CleanUpButton';
-import Tooltip from '@components/Tooltip';
+import Tooltip from '@common/Tooltip';
 import HealthIcon from '@components/Health';
 
 import Features from './Features';
